requirements.txt
annotated-types==0.6.0
anyio==4.2.0
Brotli==1.0.9
cachelib==0.9.0
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
Flask==2.2.5
Flask-Caching==2.3.0
Flask-Cors==4.0.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.7
importlib-metadata==7.0.1
itsdangerous==2.0.1
Jinja2==3.1.3
MarkupSafe==2.1.3
openai==1.25.0
packaging==24.0
pip==23.3.1
pydantic==2.5.3
pydantic_core==2.14.6
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.31.0
setuptools==69.5.1
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==2.3.8
wheel==0.43.0
zipp==3.17.0
